fix: allele become integer when setAllele

+++ main.cpp
@@ -365,7 +365,7 @@ class RealGene : public Gene {
             }
         }
 
-        void setAllele(int idx, int val) {
+        void setAllele(int idx, double val) {
             this->alleles[idx] = val;
         }
